Made changes according to review comments.
