+++ src/test/java/com/app/parallel/android/Parallel2Test.java
@@ -1,51 +1,22 @@
 package com.app.parallel.android;
 
-import io.appium.java_client.MobileDriver;
-import io.appium.java_client.MobileElement;
-import org.openqa.selenium.NotFoundException;
-import org.openqa.selenium.support.ui.FluentWait;
-import org.openqa.selenium.support.ui.Wait;
 import org.testng.annotations.Test;
-
-import java.time.Duration;
-import java.util.List;
-
-import static java.util.stream.Collectors.toList;
-import static org.testng.Assert.assertTrue;
 
 public class Parallel2Test extends BaseTest {
 
     @Test
-    public void searchWikipedia3() {
-        MobileDriver<MobileElement> driver = getMobileDriver();
-        Wait<MobileDriver<MobileElement>> wait = new FluentWait<>(driver)
-                .withTimeout(Duration.ofSeconds(10))
-                .ignoring(NotFoundException.class);
-        driver.findElementByAccessibilityId("Search Wikipedia").click();
-        MobileElement insertTextElement = wait.until(d -> d.findElementById("org.wikipedia.alpha:id/search_src_text"));
-        insertTextElement.click();
-        insertTextElement.sendKeys("BrowserStack");
-        wait.until(d -> d.findElementByClassName("android.widget.ListView").isDisplayed());
-        List<String> companyNames = driver.findElementsByClassName("android.widget.TextView")
-                .stream().map(MobileElement::getText).collect(toList());
-        assertTrue(companyNames.contains("BrowserStack"), "Company is not present in the list");
+    public void searchWikipedia4() {
+        super.searchWikipedia();
     }
 
     @Test
-    public void searchWikipedia4() {
-        MobileDriver<MobileElement> driver = getMobileDriver();
-        Wait<MobileDriver<MobileElement>> wait = new FluentWait<>(driver)
-                .withTimeout(Duration.ofSeconds(10))
-                .ignoring(NotFoundException.class);
-        driver.findElementByAccessibilityId("Search Wikipedia").click();
-        MobileElement insertTextElement = wait.until(d -> d.findElementById("org.wikipedia.alpha:id/search_src_text"));
-        insertTextElement.click();
-        insertTextElement.sendKeys("BrowserStack");
-        wait.until(d -> d.findElementByClassName("android.widget.ListView").isDisplayed());
-        List<String> companyNames = driver.findElementsByClassName("android.widget.TextView")
-                .stream().map(MobileElement::getText).collect(toList());
-        assertTrue(companyNames.contains("BrowserStack"), "Company is not present in the list");
+    public void searchWikipedia5() {
+        super.searchWikipedia();
     }
 
+    @Test
+    public void searchWikipedia6() {
+        super.searchWikipedia();
+    }
 
 }
