+++ src/test/java/com/app/parallel/ios/Parallel3Test.java
@@ -1,31 +1,22 @@
 package com.app.parallel.ios;
 
-import io.appium.java_client.MobileDriver;
-import io.appium.java_client.MobileElement;
-import org.openqa.selenium.Keys;
 import org.testng.annotations.Test;
-
-import static org.testng.Assert.assertEquals;
 
 public class Parallel3Test extends BaseTest {
 
     @Test
-    public void printText5() {
-        MobileDriver<MobileElement> driver = getMobileDriver();
-        driver.findElementByAccessibilityId("Text Button").click();
-        driver.findElementByAccessibilityId("Text Input").click();
-        driver.findElementByAccessibilityId("Text Input").sendKeys("Welcome to BrowserStack" + Keys.ENTER);
-        assertEquals(driver.findElementByAccessibilityId("Text Output").getText(),
-                "Welcome to BrowserStack", "Incorrect text");
+    public void printText7() {
+        super.printText();
     }
 
     @Test
-    public void printText6() {
-        MobileDriver<MobileElement> driver = getMobileDriver();
-        driver.findElementByAccessibilityId("Text Button").click();
-        driver.findElementByAccessibilityId("Text Input").click();
-        driver.findElementByAccessibilityId("Text Input").sendKeys("Welcome to BrowserStack" + Keys.ENTER);
-        assertEquals(driver.findElementByAccessibilityId("Text Output").getText(),
-                "Welcome to BrowserStack", "Incorrect text");
+    public void printText8() {
+        super.printText();
     }
+
+    @Test
+    public void printText9() {
+        super.printText();
+    }
+
 }
