+++ src/test/java/com/web/parallel/BaseTest.java
@@ -4,10 +4,13 @@ import com.web.BrowserDetails;
 import io.restassured.authentication.PreemptiveBasicAuthScheme;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.builder.ResponseSpecBuilder;
+import org.openqa.selenium.By;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.openqa.selenium.remote.RemoteWebDriver;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.testng.Assert;
 import org.testng.annotations.*;
 
 import java.lang.reflect.Method;
@@ -19,6 +22,9 @@ import java.util.concurrent.ThreadLocalRandom;
 
 import static io.restassured.RestAssured.*;
 import static java.util.stream.Collectors.toList;
+import static org.openqa.selenium.Keys.TAB;
+import static org.openqa.selenium.support.ui.ExpectedConditions.elementToBeClickable;
+import static org.openqa.selenium.support.ui.ExpectedConditions.presenceOfElementLocated;
 
 public class BaseTest {
 
@@ -101,6 +107,18 @@ public class BaseTest {
         js.executeScript("browserstack_executor: {\"action\": \"setSessionStatus\", \"arguments\": {\"status\": \"passed\"}}");
         driverThread.get().quit();
         driverThread.remove();
+    }
+
+    protected void testBStackDemoLogin() {
+        WebDriver driver = getWebDriver();
+        WebDriverWait wait = new WebDriverWait(driver, 10);
+        driver.get("https://bstackdemo.com");
+        wait.until(elementToBeClickable(By.id("signin"))).click();
+        wait.until(elementToBeClickable(By.cssSelector("#username input"))).sendKeys("fav_user" + TAB);
+        driver.findElement(By.cssSelector("#password input")).sendKeys("testingisfun99" + TAB);
+        driver.findElement(By.id("login-btn")).click();
+        String username = wait.until(presenceOfElementLocated(By.className("username"))).getText();
+        Assert.assertEquals(username, "fav_user", "Incorrect username");
     }
 
 }
