+++ src/test/java/com/web/parallel/Parallel2Test.java
@@ -1,41 +1,22 @@
 package com.web.parallel;
 
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.Assert;
 import org.testng.annotations.Test;
-
-import static org.openqa.selenium.Keys.TAB;
-import static org.openqa.selenium.support.ui.ExpectedConditions.elementToBeClickable;
-import static org.openqa.selenium.support.ui.ExpectedConditions.presenceOfElementLocated;
 
 public class Parallel2Test extends BaseTest {
 
     @Test
-    public void testBStackDemoLogin3() {
-        WebDriver driver = getWebDriver();
-        WebDriverWait wait = new WebDriverWait(driver, 10);
-        driver.get("https://bstackdemo.com");
-        wait.until(elementToBeClickable(By.id("signin"))).click();
-        wait.until(elementToBeClickable(By.cssSelector("#username input"))).sendKeys("fav_user" + TAB);
-        driver.findElement(By.cssSelector("#password input")).sendKeys("testingisfun99" + TAB);
-        driver.findElement(By.id("login-btn")).click();
-        String username = wait.until(presenceOfElementLocated(By.className("username"))).getText();
-        Assert.assertEquals(username, "fav_user", "Incorrect username");
+    public void testBStackDemoLogin4() {
+        super.testBStackDemoLogin();
     }
 
     @Test
-    public void testBStackDemoLogin4() {
-        WebDriver driver = getWebDriver();
-        WebDriverWait wait = new WebDriverWait(driver, 10);
-        driver.get("https://bstackdemo.com");
-        wait.until(elementToBeClickable(By.id("signin"))).click();
-        wait.until(elementToBeClickable(By.cssSelector("#username input"))).sendKeys("fav_user" + TAB);
-        driver.findElement(By.cssSelector("#password input")).sendKeys("testingisfun99" + TAB);
-        driver.findElement(By.id("login-btn")).click();
-        String username = wait.until(presenceOfElementLocated(By.className("username"))).getText();
-        Assert.assertEquals(username, "fav_user", "Incorrect username");
+    public void testBStackDemoLogin5() {
+        super.testBStackDemoLogin();
+    }
+
+    @Test
+    public void testBStackDemoLogin6() {
+        super.testBStackDemoLogin();
     }
 
 }
